// src/Organizer.SPA/src/app/api/OrganizerApiClient/model/appointmentType.ts
/**
 *  Organizer
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface AppointmentType { 
    id?: number;
    description?: string;
    billable?: boolean;
    requireCustomer?: boolean;
    color?: string;
    textColor?: string;
    shortDescription?: string;
}